changed some access modifiers in Planet according to autograder, also changed the type from int to double of the counter time variable for more accurate planet updates

+++ proj0/NBody.java
@@ -11,7 +11,7 @@ public class NBody {
 		// for(Planet planet : planetsArray){
 			// planet.draw();
 		// }
-		int counter = 0;
+		double counter = 0.0;
 		
 		while(counter < T){
 			Double[] xForcesArray = new Double[planetsArray.length];
